requirements.txt
matplotlib==3.10.1
openai==1.69.0
pandas==2.2.3
python-dotenv==1.1.0
tiktoken==0.9.0
